requirements.txt
matplotlib>=2.2.2
numpy>=1.14.2
pandas>=0.22.0
PyYAML>=3.12
ratelimit>=2.2.0
requests_oauthlib>=2.0.7
requests>=2.18.4
seaborn>=0.8.1
yamale>=1.7.0
zeep>=2.5.0